backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0
anthropic>=0.18.0
openai>=1.12.0
pypdf>=3.17.0
python-dotenv>=1.0.0
filelock>=3.13.0
